+++ airbyte-integrations/bases/base-normalization/entrypoint.sh
@@ -29,7 +29,8 @@ function configuredbt() {
     echo "Running: transform-config --config ${CONFIG_FILE} --integration-type ${INTEGRATION_TYPE} --out ${PROJECT_DIR}"
     # Generate a profiles.yml file for the selected destination/integration type
     transform-config --config "${CONFIG_FILE}" --integration-type "${INTEGRATION_TYPE}" --out "${PROJECT_DIR}"
-    # Remove config file as it might still contain sensitive credentials
+    # Remove config file as it might still contain sensitive credentials  (for example,
+    # injected OAuth Parameters should not be visible to custom docker images running custom transformation operations)
     rm "${CONFIG_FILE}"
     if [[ -n "${CATALOG_FILE}" ]]; then
       # If catalog file is provided, generate normalization models, otherwise skip it
@@ -57,7 +58,8 @@ function configuredbt() {
     # Generate a profiles.yml file for the selected destination/integration type
     echo "Running: transform-config --config ${CONFIG_FILE} --integration-type ${INTEGRATION_TYPE} --out ${PROJECT_DIR}"
     transform-config --config "${CONFIG_FILE}" --integration-type "${INTEGRATION_TYPE}" --out "${PROJECT_DIR}"
-    # Remove config file as it might still contain sensitive credentials
+    # Remove config file as it might still contain sensitive credentials  (for example,
+    # injected OAuth Parameters should not be visible to custom docker images running custom transformation operations)
     rm "${CONFIG_FILE}"
   fi
 }